requirements.txt
pymongo==4.6.0
pyyaml==6.0.1
requests==2.31.0